feat(interface): Change API fetching to data passing
Instead of passing the API endpoint to get the data from
the component now follows a more convenient format of passing
the data into it with a dataState prop on the main component.
The setData part of the data state enables the component to
modify some field which are required for the interactive
legend. Closes #24.